Unix vWii patch support

+++ UNIX/patch.sh
@@ -35,19 +35,16 @@ while true; do
   esac
 done
 
-if [ "$TYPE" = "v" ]; then
-	printColor '1. Looking for WAD...'
-  # Users of vWii must provide their own wad.
-  if [ ! -f ./IOS31_v3864.wad ]; then
-    echo "Warning! I'm missing the actual IOS, named IOS31_v3864.wad."
-    echo "I can't tell you how to obtain this, but please place it in the current directory under that name."
-    exit 1
-  fi
-else
+if [ "$TYPE" = "w" ]; then
   # For Wii
   printColor '1. Downloading IOS from NUS...'
   mono Sharpii.exe NUSD -ios 31 -v latest -o IOS31-old
   mono Sharpii.exe NUSD -ios 80 -v latest -o IOS80-old
+elif [ "$TYPE" = "v" ]; then
+  # For vWii
+  printColor '1. Downloading IOS from NUS...'
+  mono Sharpii.exe NUSD -ios 31 -v latest -o IOS31-tik
+  mono Sharpii.exe NUSD -id 000000070000001F -v latest -o IOS31-old
 fi
 
 printColor '2. Unpacking IOS...'
@@ -56,30 +53,47 @@ if [ "$TYPE" = "w" ]; then
   cp IOS80-old/0000000100000050v6944.wad IOS80-old.wad
   mono Sharpii.exe WAD -u IOS31-old.wad IOS31/
   mono Sharpii.exe WAD -u IOS80-old.wad IOS80/
-else
-  mono Sharpii.exe WAD -u IOS31_v3864.wad IOS31/
+elif [ "$TYPE" = "v" ]; then
+  cp IOS31-old/000000010000001fv3864.wad IOS31-old.wad
+  mono Sharpii.exe WAD -u IOS31-old.wad IOS31/
+  cp IOS31-tik/cetk IOS31/000000010000001f.tik
 fi
 printColor  '3. Patching the APP file...'
-xdelta3 -f -d -s ./IOS31/00000006.app ../Patches/"$TYPE"00000006-31.delta ./IOS31/00000006.app
+
+if [ "$TYPE" = "v" ]; then
+  VWII="-vwii"
+else
+  VWII=""
+fi
+
+xdelta3 -f -d -s ./IOS31/00000006.app ./00000006-31"$VWII".delta ./IOS31/00000006.app
 # We only do IOS80 for actual Wiis.
 if [ "$TYPE" = "w" ]; then
-  xdelta3 -f -d -s ./IOS80/00000006.app ../Patches/"$TYPE"00000006-80.delta ./IOS80/00000006.app
+  xdelta3 -f -d -s ./IOS80/00000006.app ./00000006-80.delta ./IOS80/00000006.app
 fi
 printColor '4. Packing the new IOS...'
 mkdir WAD
 # This verion of Sharpii has fakesigning enabled by default.
-mono Sharpii.exe WAD -p IOS31/ ./WAD/IOS31.wad
+mono Sharpii.exe WAD -p IOS31/ ./WAD/IOS31.wad -f
 mono Sharpii.exe IOS ./WAD/IOS31.wad -fs -es -np -vp
 
 if [ "$TYPE" = "w" ]; then
-  mono Sharpii.exe WAD -p IOS80/ ./WAD/IOS80.wad
+  mono Sharpii.exe WAD -p IOS80/ ./WAD/IOS80.wad -f
   mono Sharpii.exe IOS ./WAD/IOS80.wad -fs -es -np -vp
 fi
 printColor '5. Cleaning up files... (Type Y if needed).'
 rm IOS31-old.wad
-rm IOS80-old.wad
 rm -rf IOS31-old
-rm -rf IOS80-old
 rm -rf IOS31
-rm -rf IOS80
+
+if [ "$TYPE" = "v" ]; then
+  rm -rf IOS31-tik
+fi
+
+if [ "$TYPE" = "w" ]; then
+  rm IOS80-old.wad
+  rm -rf IOS80-old
+  rm -rf IOS80
+fi
+
 echo "We're done."
